split db.py and add sort_competition
