+++ gudlft/.env
@@ -1,2 +1,2 @@
-TESTING=False
+TESTING=True
 HEADLESS=False
