+++ gudlft/json/test2_competitions.json
@@ -9,6 +9,16 @@
             "name": "Test competition2",
             "date": "2023-08-09 00:00:00",
             "numberOfPlaces": "20"
+        },
+        {
+            "name": "Test competition3",
+            "date": "2023-10-09 00:00:00",
+            "numberOfPlaces": "20"
+        },
+        {
+            "name": "Test competition4",
+            "date": "2020-08-09 00:00:00",
+            "numberOfPlaces": "20"
         }
     ]
 }
