+++ gudlft/server.py
@@ -4,7 +4,7 @@ from distutils.util import strtobool
 
 from flask import Flask, render_template, request, redirect, flash, url_for
 
-from .db import DataLoader
+from .models.dataloader import DataLoader
 
 from dotenv import load_dotenv
 
@@ -47,7 +47,7 @@ def showSummary():
                 'welcome.html',
                 club=club,
                 competitions=data.competitions
-                ) 
+                )
     except IndexError:
         flash("!!! Cette adresse email n'est pas reconnue")
         return render_template(
